[OpenXR] Fix VR180 stereo left to right videos
With the transition from OculusVR to OpenXR some of the VR video
formats started to fail. Some time ago we fixed the 3D side-by-side
by modifying the code in VRVideo that creates the proper projection
layer to render the video.

Now is the time for VR180 stereo from left to right. This commit
adds an alternative implementation of create180LRProjectionLayer()
that does not require a equirect for each eye as in the case of
OculusVR. With OpenXR we can just create one and assign a different
rectangle to be shown on each eye.

In addition it also removes an unneeded extra matrix multiplication
from the equirect OpenXRLayers code.

Fixes #562