Do not consider YouTube's 3D audio videos as immersive
Our YouTube extension contains some heuristics that try to detect
the right VR projection to be used in the videos depending on the
text in the description. They are heuristics so indeed not perfect.
One of the ways to improve them is by not considering 3D audio
videos as immersive videos.

In general it's better to have false negatives than false positives
because users can switch from non-immersive to immersive projections
but not the other way around. So we should be very conservative
when automatically switching to immersive.