Fix the destructor of OpenXRLayerBase
The URL bar could become unusable after opening and closing windows,
and also when switching between regular and private mode.

The same could also happen to the top bar buttons.

The cause of the bug was that the default destructor of
the OpenXRLayerBase class was calling the Destroy() method.

That change had been added in commit 2b3eb21
https://github.com/Igalia/wolvic/pull/554

Fixes #582