Fix package name for mainland China
For mainland China, the package name is com.cn.igalia.wolvic, which
is different from the one for overseas. We do have code that
automatically does the replacement when generating the package.
However the recent changes in the build system that added more flavours
broke it.

Adjusted the pattern used to detect HVR builds for mainland China so
that it properly replaces the package name automatically.