Upgrade the androidx.work version
Starting from API version 31, PendingIntents must be created
using either the MUTABLE or the IMMUTABLE flag. Not doing that
causes a crash. We fixed that inside Wolvic, but some of the
libraries we were using were too old, and thus were not passing
the flags yet.

Upgraded the version to 2.7 which is the first one shipping
the fix.