[Spaces] Do not link against qxrclients from Unity SDK
We have been linking against both the OpenXR loader and the qxrclients
packages distributed in the SDK for Unity. That's fine for the loader,
however in the case of qxrclients, the package distributed is old
and was actually causing hand tracking not to work properly in Wolvic.

Removing that dependency makes hand tracking work. The *aar file will
be also removed from the third_parties repo as it is no longer needed.