Remove specific <31 targetSDKs required for the PendingIntent crash
We were using API version 30 for some of the platforms because
they crashed with 31 due to the PendingIntent change (the
MUTABLE/IMMUTABLE flag is now mandatory). Now that this is no
longer a problem after the update of the androidx.work version,
we can safely remove them and use the common one (which
currently targets 31).