Remove the Menu hand gesture
When hand tracking is active, touching your thumb and ring fingers
was being mapped to a click of the Menu button in the controller.
In turn, this click was interpreted as a request to navigate Back.

The consequence of the sequence above was that many common actions
would trigger this gesture and cause an unwanted Back navigation.

One case where this happened often was simply picking up the controller.

This PR removes that Menu gesture.