Make OpenXRLayerPassthrough an actual OpenXRLayer
The current implementation of OpenXRLayerPassthrough does not fit
well with the OpenXRLayer architecture because it does not inherit
from it. This causes several issues, among others there are explicit
calls in the code to Request/Clear draw of the layer.

The reason behind that is that there were some issues with the
structure defined by the passthrough extension, we're talking about
XrCompositionLayerPassthrough:
1) It does not have eyeVisibility as other composition layers
2) The layer flags are stored in an attribute called flags instead
of layerFlags as XrCompositionLayerBaseHeader defines.

The first issue was addressed in PR#620. We're fixing the second one
as part of this. The trick is simple, we just need to access the
structure using a XrCompositionLayerBaseHeader cast. The passthrough
struct is binary compatible, it's just that they decided to use
a different attribute name (flags vs layerFlags), but the types and
the location of the attributes is identical.

After doing that, we can safely make OpenXRLayerPassthrough extend
from OpenXRLayerBase. The Init, Update, etc... methods now do override
OpenXRLayerBase's. The benefit is that we don't need to explicitly
now request/clear draws as that is automatically done by the current
machinery. We just had to create a new rootnode that will add
the passthrough layer to the drawList whenever needed.

Last but not least, the creation of the passthrough layer is now
done in TogglePassthrough where it does make much more sense than
in StartFrame().