Initialize eyeVisibility in OpenXRLayerBase subclasses
So far we have been doing it at OpenXRLayerBase level because
all of our composition layers had eyeVisibility field. However
that assumption is not completely correct, for example the
PassthroughFB layer does not have it. We must only assume
that they are XrCommpositionLayerBaseHeader (which does not
have eyeVisibility).

That's why we are moving the eyeVisibility initialization to
the different OpenXRLayer subclasses supporting it.

For the same reason HeaderCount() returns now GetNumXRLayers()
directly as the latter does not depend on OpenXR structures
but our own data stored in our layers.