[OpenXR] Add support for passthrough strategies
There are multiple ways to perform passthrough depending on the
platform. Some of them might use a compositor layer (like Meta)
while some others are inherently made of mixed reality and just
need an specific blend mode (like Lynx).

A strategy pattern is used to deal with those different behaviours
isolating them from the rest of the code. From now on the
DeviceDelegate will store whether or not passthrough is active.