[OpenXR] Add blend mode passthrough support and enable it for Lynx
After generalizing the code that performs the passthrough a bit, we
go further and move it completely to the OpenXRPassthroughStrategy
classes. The different strategies allow us to do passthrough either
by using a compositor layer or via changing the environment blend
mode.

Almost all the passthrough code is now isolated from the DeviceDelegate,
basically only the passthrough layer was kept there for coherence reasons
(other optional layers like the video layer are also there).

After everything is in place we can safely enable passthrough by
default for Lynx as we've already verified that the blend mode
strategy works fine there.