New back-navigation gesture using index-pinch when left hand faces head
Current gesture for navigating back uses a pinch geature between thumb
and ring fingertip. This gesture is far too easy to trigger involuntarily
and results in many usability issues (and user complains).

While we are still defining a comprehensive solution for all hand-interactions
in Wolvic, this change serves as a short/mid term fix. It replaces the
existing gesture by a new one: a normal pinch (thumb + index fingertip)
with left hand facing the head. This is inspired by Oculus behavior, where
a hamburger icon is shown when left hand is facing head. This specific
icon is already generalized on the controllers to produce a navigate-back
action.

A circled, hamburger menu icon is shown locked to the index's fingertip,
when the left hand is facing the head. As the user performs a thumb+index
pinch, the icon reduces its size up to 50%, when the pinch is closed
and a navigate-back action is triggered.

It has been tested on Quest2 and Pico4 devices.

In the future, we could generalize this new floating+button+with+icon
as a new widget that reacts to pinching with specific fingers, and it
will probably work for the current Pointer widget, which would become
a variant of it.