Add 1€ filtering for hand tracking aim
Hand tracking data carries a lot of noise. That's why if we use
the raw pose data from hand joints to define an aim (pointer ray)
we get a cursor with a lot of jitter. In order to fix that
we use the 1€ filter, which is a simple lightweight filter specially
designed for this kind of data coming from the human body.

See https://gery.casiez.net/1euro/ for additional details.