[Spaces] Workaround incorrect stage transform
Same as with HVR the SnapdragonSpaces runtime does not provide
a valid stage transform, and thus any WebXR experience is
rendered at floor level. We can workaround that by hardcoding
a specific translation on the Y axis. We do use a similar
value provided by other runtimes which correctly implement it.