[Spaces] Runtime incorrectly always report joint locations as inactive
The current SnapdragonSpaces runtime always report hand joint locations
as inactive in the XrHandJointLocationsEXT, meaning that they are
not being tracked. Workaround it by considering that they are always
tracked for Spaces. This won't detect the cases of them not being
tracked but at least it works in the most common case.