[Spaces] Workaround hand joint locationFlags are always 0
The current SnapdragonSpaces runtime (0.12.1) does not properly
fill in the locationFlags. They're always 0, meaning that we
don't have information about whether a pose is valid/invalid
or even if it's tracked at all.

We can workaround it because we found out that when hand joints
are not tracked the system returns (0,0,0) for the position.
Obviously we won't detect that case anymore but it's highly
unlikely to happen.