[Chromium] Refactored GeckoSurfaceTexture
This class was called GeckoSurfaceTexture but it does not really
depend on Gecko more than a couple of class names. The rest of the
file is about making JNI calls. Provided that we're inheriting
the same mechanism for WebXR content in the Chromium backend,
we can simply reuse this class replacing the actual names of
the classes/methods.

As part of that refactoring we're renaming it to
EngineSurfaceTexture to make it clear that is engine agnostic.