Workaround for YouTube videos
Recent changes in the YouTube page have caused that we no longer
receive media playback events, which we use to recognize when a video
is playing, etc.

In particular, this affects videos which require a special projection
(stereo, 360 degrees, etc.). Since we can't detect the playing video,
we don't offer the possibility of using the right projection for it.

This workaround makes it possible to access the media projection
menu from a fullscreen YouTube video even if we don't have access
to its information and playback events.

Most of the changes are located in NavigationBarWidget.java

Additionally, AnimationHelper will avoid flickering when trying to
show/hide a widget that is already visible/hidden.