[Regression] Fix background dimming on fullscreen videos (#647)
In 0ef55bc7 we added a new method called PushNextXrStructureInChain
to easily append structures to the "next" pointer on an structure
chain. This kind of structures are tipically used to setup extensions.

We did that, among other things, to support the CompositionLayerColorScaleBias
extension that allowed us to do background diming on videos in OpenXR
as we used to do with OculusVR backend. The problem is that we
didn't pass the right pointer to that method so that the structure
was not really added to where it should.