Bump version code
This was done after releasing 1.3.4 for all the platforms just for
Pico. We found a crash caused by some lack of permissions, and since
the packages were still under review, we decided to withdraw the
submission and generate new APKs with the crash fix. This does not
deserve a different release, so it should account as an exception
in the release process.