Move the UI control of the passthrough to the VRBrowserActivity
The mIsPassthorughEnabled flag is now defined in the VRBrowserActivity and
and new function has been defined in the WidgetManagerDelegate interface, so
that the HamburgerMenuWidget can access to it in order to set the proper
icon in the menu item.

Additionally, the code to enable the passthrough by default in certain
runtimes has been also moved to the VRBrowserActivity class.