Set the correct trigger and squeeze button states when hands disappear
Since we introduce the ability to emulate controllers using the hands, with
thumb+index pinch as trigger and thumb+middle pinch as squeeze, we never checked
that these hand joints had valid position data, to properly calculate the
distance between them.

This is causing that when hands are occluded or completely disappear, the
joint data becomes invalid and we may incorrectly detect a pinch gesture. This
results in a hand that is not active or visible, could lock the trigger or
squeeze buttons in a pressed state, preventing the other hand or the controllers
from getting the focus.

This patch makes sure that we detect a pinch gesture when the actual joints
have valid positions, and always set the trigger and squeeze button states.

Fixes #655