Properly activate back action in non Meta devices
There was some code to activate the back action for non Meta devices
that was based on Pico behaviour. However that is not completely
correct because it was based on the Pico runtime which has a lot of
issues when reporting orientations of the hand joints. Among others,
the axis for the palm joint are interchanged.

In order to make it work for a variety of devices we change the code
that checks whether the palm and head are looking at each other.
Basically compute the dot product (angle) of two vectors originating
from head and palm. If that value is higher than a threshold (0.7)
 we consider them facing the same direction, i.e., palm and head
looking at each other. Note that negative values mean opposite
directions.