Disable aim state during back gesture
We should disable the aim state whenever the back action is performed,
i.e., whenever the left hand faces head. This does work correctly
for platforms using the FB aim extension because in those cases the
mHasAimState is set to false. However we were not doing that for
all the configurations not using the extension. The gesture was
properly detected but we were still reporting a valid aim, and thus,
the ray was still painted.

The code has been generalized a bit, so that it also works in systems
that do not support the FB aim extension, but that do take over the
hand tracking whenever the hands face the head.

Fixes #678, fixes #672