Fix ability to trigger a pressed event during hand-tracking when no aim is enabled
This is a bug that was introduced recently while trying to fix a
different issue (see PR#670 <https://github.com/Igalia/wolvic/pull/670>).

The problem is that we are not considering the aim state when deciding
when a pinch gesture should trigger or not a button press or a select
action.