New extension-less method to compute hand tracking aim
So far we've been computing the hands aim using a FB extension.
It works pretty well but it is not supported in many devices.
That's why we add a new method to compute the aim, based on
some research from Ultraleap
 https://docs.ultraleap.com/ultralab/far-field-ray-blog.html

Basically we construct an aim ray from two points, one located
in the shoulder and the other one located in the wrist.
We get the wrist pose from the XR_HAND_JOINT_WRIST_EXT joint.
The shoulder pose is then inferred from the position of the
head (using average proportions of an adult human body).