Update non_L10n.xml
Changes the url of the feedback form so that it works (it changed and is now broken entirely).  I think this should probably lack the /en/ part but that currently doesn't actually work either so maybe we can at least merge it into parity with where it was working as of earlier today.