Declare Wolvic as a browser application in Android
This allows Wolvic to be listed in the Android's
"Settings->Apps->Default Apps->Browser app" menu. Once there users
can select it to be the default browser of the system to handle
http and https schemas.

This has been broken for a while since we started to handle the
special wolvic:// URLs.