Get the device model from Android OS instead of from OpenXR runtime
The values returned by the OpenXR runtimes are generally not reliable.
In many cases the same names are used for different devices by mistake
while in some others the returned name does not even identify a device
(it's generally the runtime name). Perhaps that's intended (after all
OpenXR aims for device independence) but we need a way to identify
specific devices (for example to render the proper 3D models for
the controllers).

This new method allows us to remove also some workarounds we had to
set in place to properly detect the Meta Quest Pro for example.