Remove the OpenXR based system filter
We used to check the system name returned by the OpenXR runtime to
filter the list of the supported interaction profiles. However that
was not very reliable mainly because the value returned by the
runtimes was wrong in many of the cases (duplicates, empty values,
changing names, wrong device names...).

Now that we retrieve the device type from Android OS, we can replace
that filtering by another one based on the device type. That allows
us also to get rid of a workaround we had to add for older versions
of the Pico OS (they changed the name of the devices in between
releases).

Some of our supported devices are not really supported on the Gecko
side, meaning that their proper controller models won't be rendered
in WebXR sessions. We basically have to pretend to be other
supported devices to get controller support.