Make passthrough work on release builds in Meta devices
Passthrough support was added to meta some time ago using a FB
extension. That extension required the creation of a
XrCompositionLayerPassthroughFB that was passed to the compositor.
That was working for debug builds but not for release ones. The reason
why is because we were storing that object in a local variable that was
freed after exiting its scope. It was not enough adding them to a vector
because we're using a typecast for the compositor layer generic header.

Instead we now store that composition layer in the OpenXRLayerPassthrough.
That requires us to implement the Update() method to initialize it
properly before passing it to the compositor. That makes our passthrough
layer identical to the other ones (as we were not using/overriding Update()).

Fixes #703