[AOSP] Added the build machinery for generic AOSP devices
It builds the android loader from the official OpenXR-SDK repo
which is now in the third_party repo. As the build fails due to
us using -Werror, we have to patch it locally for the build.

Also enable x86_64 ABI for AOSP since there are devices that are
not ARM like the MagicLeap2

This includes the addition of some required files, mainly
* Android manifest
* Interstitial widget
* PlatformActivity