Avoid creating hand models renderer object when not used
Right now we are creating the HandModelsRenderer instance in all cases,
even on devices that don't support hand-tracking at all, or where we
are rendering joint spheres instead.

This patch makes HandModelsRenderer object be created lazily, right before
rendering the hand models. At this point we already know that the device
supports hand-tracking, and a hand mesh is enabled and already loaded for it.

The patch also removes HandModelRenderer class' inheritance from
from vrb::ResourceGL. The reason is that on ResourceGL objects, the OpenGL
state is not initialized immediately, so it doesn't play well with us
creating HandModelsRenderer lazily.

Fixes #701 <https://github.com/Igalia/wolvic/issues/701>