[OpenXR] Query&select environment blend modes from the runtime
Enumerate the available environment blend modes in the runtime
and select the proper ones for both default operation and
passthrough modes. The preferred selections for the default
mode are: OPAQUE -> ADDITIVE -> ALPHA

For passthrough mode we do ADDITIVE -> ALPHA -> OPAQUE. If
OPAQUE then passthrough might still be supported, for example
via an OpenXR extension (as it happens with Meta devices).

Now that blend modes are automatically selected from the supported
blend modes retrieved from the runtime, we can get rid of the code
that returned the blend mode in the passthrough strategies as it is
no longer needed.

This allows us also to remove the NoSkybox strategy which was
actually not needed (if passthrough is enabled and there is no
passthrough layer then we simply don't draw the skybox).