Fix auto-enable passthrough for the Lenovo A3
The Lenovo A3 should use passthrough mode by default as it's a
see-through device. We tried to toggle passthrough on app start
but that's in general too early because the device type might
not have been set. It's better to do it just after getting the
proper device type from native code.