[ML2] Adding MagicLeap 2 controller support
We're adding the OpenXR interaction profile and all the required code
to make the controller work. Note that contrary to other devices,
the interaction profile requires a vendor extension to be active in
order to work.

Also we are not including any 3D model of the controller on purpouse
because as the device is see-through the real controller can be
always seen so there is no point on rendering a 3D model on top.