Align windows vertically
To improve usability and legibility, browser windows will be centered vertically
relative to their default position.

The way to implement this is to add a "parentAnchorGravity" field to widget
placements, which can be used to center widgets relative to their parent.

Browser windows to the left and right of the user are positioned relative
to the central window. Therefore, when that central window changes its height
we need to reposition it so it stays vertically centered.

Center windows can be enabled and disabled in the Display settings.