Fix zoom in desktop mode
Disabling apz.allow_zooming was not only disabling zooming after a
pinch gesture but also in response to viewport size changes as when
we resize the window. That does not seem coherent with the API
docs for that setting but that's what happens.

It seems that we can effectively remove the zoom effect by setting
an empty command for the various phases of the pinch gesture.