Add appInChina store
In the past we had a dimension for HVR packages for mainland China
and overseas. We're now going to use the services of AppInChina to
publish Wolvic there, so we can remove the "country" dimension and
convert the "cn" value into a "appInChina" value for the store
dimension. That would allow us to have a separate build configuration
for the Chinese market for any of the existing configurations.

This allows us to slightly simplify the handling of the packages
for China since some of the configurations won't be specific to HVR.