Compute a better aim ray for hand tracking
The XR_FB_hand_tracking_aim extension is used to compute the aim
(i.e. the pointer) used when interacting with Wolvic via hand
tracking. It works pretty well but it isn't supported in many
platforms.

For the other systems we were using the orientation quaternion
from a hand joint (either the palm or the wrist). The aim
resulting from that is quite unstable and too sensitive to the
movements of the hand. What's more, clicking on a specific spot
is rather difficult because the pinching action does indirectly
modify the orientation, and thus, the location of the pointer.

This commit implements a different approach. First we estimate the
position of each shoulder based on the location of the head.
Then we trace a ray from that estimated position to a joint
location in the hand. The resulting ray is much more stable.
The joint location that we selected is the middle knuckle because
it provides the stability of the wrist (for example for pinching)
while at the same time allows to fine tune the position of the
pointer by slightly rotating the wrist.