Show a confirm prompt when trying to exit the app via back action
In non-immersive mode, triggering the back action when there is no
browser history left should launch a confirmation dialog to exit
the application. This was only available for HVR but should be
actually enabled for all the platforms.

Also call finishAndRemoveTask() in order to gracefuly close the
application after clicking on the close button in the dialog.