Cleanup the native code finalization process
This cleanup involves various steps:
* Use aAppState->destroyRequested==0 as the run condition for the
main render loop. This allows us to handle the same way the cases
in which the system closes the app and the cases in which the app
tries to exit itself.
* Process OpenXR events before drawing. This way we can early
detect whether the session is being finalized and not to draw in
those cases.
* Call ANativeActivity_finish() if the session is being finalized.
By doing this aAppState->destroyRequested will become !=0 and the
main render loop will exit.
* Move all the finalization code just after the main render loop
instead of calling it and then early return.
* Call xrRequestExitSession() whenever device->ExitApp() is called.
This allows us to properly finalize the OpenXR session. This
makes us a better OpenXR citizen.