Do not call ProcessEvents() inside StartFrame()
This used to fit well in the OculusVR plaftform but is problematic
(and not needed) for the OpenXR backend. For example this sequence:

1. User tries to close the app -> xrRequestExitSession() is called
2. The render loop calls Draw() before the runtime is able to exit
the current session. The Draw() calls BrowserWorld::StartFrame()
3. StartFrame() calls device->ProcessEvents
4. We receive a XR_TYPE_EVENT_DATA_SESSION_STATE_CHANGED event
announcing that the session is STOPPING so vrReady is set to false.
5. The StartFrame() call continues and at some point
DeviceDelegateOpenXR::StartFrame() is called with vrReady==false
which is an error condition.