[OpenXR] Add system gesture recognition to GestureManager
Many devices trigger some system event whenever the palms
face the head. For example Meta takes control over the hands
and show a menu. Pico also starts a timer to launch a menu.
We do use that gesture also to trigger navigation back or
to exit the application in other devices.

This code that was in OpenXRInputSource is properly placed
now in the GestureManager so that it does not interfere
with the rest of the generic code for controllers. This
allowed us to remove some attributes from OpenXRInputSource
that really belong to the GestureManager.