Add support for XR_MSFT_hand_tracking_mesh extension
Will be enabled on all devices that supports it, and has been
tested on the Lenovo A3.

This extension feed applications with hand mesh geometry on a per-frame basis,
with all transforms already applied to all vertices and normals. This differs
significantly from our current hand mesh rendering, which uses a skinned mesh
and only the skeleton (joints) are updated per-frame.

A new class `HandMeshRendererGeometry` is created to encapsulate this hand
mesh rendering mechanism.