Fix duplicated toolbars in phone UI
Sometimes pages in the phone UI have two toolbars, which looks
like a visual glitch.

This was happening because fragment_content_holder.xml had a toolbar,
but at the same time it was embedding other fragments that also
had their own toolbar.

This PR cleans that up and ensures that the layout is always correct.