Support per-architecture dependency substitutions for Gecko
MagicLeap2 is x86, so it needs an x86 Android gecko to build.
With this PR, we don't have to manually change the location
of the gecko build dir to use it.

If you want to build for different architectures at the same
time, you can specify the aarch64 path via
`dependencySubstitutions.geckoviewTopobjdir` while the x86_64
path via `dependencySubstitutions.geckoviewTopobjdirX64` in
local.properties

Signed-off-by: Songlin Jiang <sjiang@igalia.com>