Fix left hand action button on Lynx (#842)
When we introduced the generic hand mesh rendering we broke the left
hand button action on Lynx. This patch fixes that and also simplifies
the workflow of storing and indexing into the joint matrices vector.

The issue is that the skeleton of our left hand model has the order
of the joints in reverse order with respect to what the
XR_EXT_hand_tracking extension gives. So we had to correct the order
of the matrices there. The problem is that we were doing that too early,
during assignment and before calculating the left hand action button
matrix, which requires indexing the joint matrices to get the thumb
and index tips.

Both Oculus and Lynx use the generic mesh rendering code-path, but
only Lynx uses the left hand action button (in Oculus it is already
provided by the system). Hence the bug went unnoticed until we tested
it on Lynx.

Fixes #839 <https://github.com/Igalia/wolvic/issues/839>