Support using maven GV from the release channel (#882)
We supported it at some point but it was broken for some time and
we eventually removed it. Now we can restore it. Release versions
of Wolvic must not use maven GeckoView versions because they don't
include our updates to the Wolvic-Gecko communication protocol,
meaning that WebXR won't work.

However we can still use it for super-quick bisects for example,
or to test 2D browser features in advance without having to
build Gecko.

Co-authored-by: Hollow Man <hollowman@opensuse.org>