[Chromium] Mute tabs when inactive
In 50d4e199 we added support for notifying the web contents about
the visible status of some tabs. This allowed us to pause some
media playback when the tabs go to background. However that is not
enough for all media content. For that we need to call a couple of
methods: suspendAllMediaPlayers() and setAudioMuted() to ensure
that no playback is performed.