Use per web engine URI overrides
Many services in the Web assume that the Meta Browser (Chromium based)
is the only browser supporting WebXR content. That's why we
sometimes need to spoof the User Agent and pretend to be what we
are not.

This is very specific to the web engine though, for example
it should not be required for the chromium backend as we
use a chromium UA. That's why we're moving the list of URI
overrides to web engine specific directories.