Increment timeout when waiting for frame
Right now we are using a timed wait of ~100ms when waiting for
frame completion in gecko. Some heavy WebXR experiences can easily
hit this limit, specially when loading resources, causing the
frame to be discarded due to timeout. This causes unnecessary
stutter, which can be avoided if we wait a bit more for the
frame, at the expense of the frame rate.

This patch increases the frame waiting timeout to ~250ms.