Fix a crash when drawing hands before updating mesh in HandMeshRendererSkinned
The order of calls to Update() and Draw() on a hand mesh renderer is
controller by BrowserWorld, and is not deterministic. To protect
against trying to Draw() before Update() has been called, we bail
early on the Draw() call if there are no hand joint transforms.

Fixes https://github.com/Igalia/wolvic/issues/934