Properly emulate controller primary action with hand tracking
So far the value of the primary action button when emulating the
controller was permanently set to 1.0. That's wrong, we should
use the computed distance between the pinching fingers instead.

Apart from that we're also fixing the value of touched. We should
consider that it's "touched" whenever the pinching action starts.

Fixes #480