[Chromium] Add null checks for MediaSessionDelegate access
The first thing the session does when closing a session is cleaning
up the session listeners. This involves, among other things, setting
all the delegates to null.

It's still possible to get media events from the web engine after we
proceed to close the session (like inactive notifications) so we must
ensure that we don't run any job in that case and ignore the notification.

We went the extra mile and we've also added null checks for a few
more insecure usages of getMediaSessionDelegate().

Fixes #915