Fix restarting Wolvic
Resolves #949

Based on the answer from https://stackoverflow.com/a/71392776

This works with Android > 10, tested up to Android 13 (preview).

Someone said that they also tested in Android 7 and it works well.

Since Android 10, you can't launch from the background, so the old
method is no longer working:
https://developer.android.com/guide/components/activities/background-starts

The new one no longer needs to schedule a start from the background,
so just remove the scheduled way of restarting.

Signed-off-by: Songlin Jiang <sjiang@igalia.com>