[Chromium] Fix a startup crash
The recently added SessionFinder that looks for text in the page
does not have implementation for the chromium backend. The current
code does not handle well the case of finder being a null pointer
as it is not supposed to happen. Added a temporary workaround
to keep the chromium port working until we don't implement it.