Stop having findInPage and Panel opened at same time
We shouldn't let user feel that we can search content
in the library page through findInPage. So we close
findInPage when we open the Library, and we close the
panel when we open findInPage in the hamburger menu.

We shall pay attention to this as well in
https://github.com/Igalia/wolvic/pull/881
once this is merged

Signed-off-by: Songlin Jiang <sjiang@igalia.com>