Allow open new page without interrupting video playing
If there's no video playing in the current window, we still
open the new page in a new tab.

If we have video playing in current window, ensure we don't
open a new tab in a window that has active video, but open it
in the Window that has no video playing.

If all the current opened Windows have video playing, we have
to open uri in a new window. If we have maximum window number,
then open the uri as a new tab in current window.

Continue addressing #940

Related to #958

Signed-off-by: Songlin Jiang <sjiang@igalia.com>