[OpenXR] Cleanup swapchain destruction
Swapchains must be destroyed before destroying the XrInstance.
We can find them in the views and in multiple UI layers, including
quads, cylinders, skybox or passthrough layers. We were relying on
smart pointers to properly destruct the layers. That was happening,
however it was done too late, i.e., after destroying the XrInstance.

That's why we explicitly have to destroy the layers before destroying
the XrInstance so that they in turn are able to destroy the swapchains
on time.

This fixes a long lasting error on app shutdown:
Error [GENERAL | xrDestroySwapchain | OpenXR-Loader] : No active XrInstance handle.

Fixes #546