Refactor CI for external contributors
Now we split the CI into internal and external.
- The external one doesn't need the secret (doesn't need third-party
SDKs by building the noAPI flavor). So we can quickly test the PRs from
external contributors.
- The internal CI works by checking if the secret is present in the
repository. If it is, we then continue running the build part.

Now we also allowing running CI when pushing into the branches, so that
we can have a full test on the PRs once the PRs are merged.

Signed-off-by: Songlin Jiang <sjiang@igalia.com>