Upgrade R8 to version 8.2
R8 is the tool used to minimize (shrink + optimize) the Java code.
We normally just run it for release packages as it slows down the
build significantly. Whenever the Chromium backend is used, that
minimization was failing because R8 was unable to parse some
proguard rules provided by Chromium.

The problem is that the code that parses those rules was added in
R8 8.2. The latest available version in the maven repositories
we use is 8.1, so we have to add a new repo to get that.