[NoAPI] Initialize Java in VRBrowser before creating the DeviceDelegate
This allows Wolvic to properly create the external context that will
be passed to the Web engine. Without that external context the web
engine is not able to create the shared memory that will be used
to talk to Wolvic, and thus no WebXR would be possible.

This allows NoAPI flavour to enter WebXR experiences.