[Chromium] Record history
GeckoSession uses a HistoryDelegate object (provided by Wolvic) to
keep track of history changes. In the case of Chromium the concept
is different. We can listen to the events dispatched by the
WebContents object and manyally call the HistoryDelegate to record
changes in the history. That way we don't need to change anything
in the Gecko port and we get history support in Chromium.