Improve detection of URLs with long gTLDs
When some text is typed in the URL bar we try to detect whether it's just
plain text to be passed to the search engine, an special page (like about:x)
or an URL. The parsing code used to detect valid URLs was considering that
a gTLD could only have up to 5 characters. That was fine for the original
gTLDs but has not been true for some years. There are gTLDs with up to
24 characters even.

Fixes #990